{"schema":"decodex/commit/1","summary":"Modularize app server runtime internals","authority":"XY-1115"}
